:lock: Apply guard to user controller

+++ src/users/users.controller.ts
@@ -2,11 +2,15 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Patch,
   Post,
+  Req,
+  UseGuards,
 } from "@nestjs/common";
+import { JwtAuthGuard } from "auth/guards/jwt-auth.guard";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
 import { UsersService } from "./users.service";
@@ -25,13 +29,27 @@ export class UsersController {
     return this.usersService.create(createUserDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Patch(":id")
-  async update(@Param("id") id: string, @Body() updateUserDto: UpdateUserDto) {
+  async update(
+    @Req() req,
+    @Param("id") id: string,
+    @Body() updateUserDto: UpdateUserDto,
+  ) {
+    const curUserId = req.user.id;
+    if (curUserId !== id) {
+      throw new ForbiddenException("无权更新");
+    }
     return this.usersService.update(id, updateUserDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(":id")
-  async remove(@Param("id") id: string) {
+  async remove(@Req() req, @Param("id") id: string) {
+    const curUserId = req.user.id;
+    if (curUserId !== id) {
+      throw new ForbiddenException("无权删除");
+    }
     return this.usersService.remove(id);
   }
 }
